+++ crates/cli/src/commands/tool/info.rs
@@ -1,22 +1,35 @@
 use crate::printer::Printer;
 use clap::Args;
-use proto_core::{detect_version, load_tool, Id, PluginLocator};
+use miette::IntoDiagnostic;
+use proto_core::{detect_version, load_tool, ExecutableLocation, Id, PluginLocator};
+use proto_pdk_api::ToolMetadataOutput;
 use serde::Serialize;
 use starbase::system;
 use starbase_styles::color;
+use starbase_utils::json;
+use std::path::PathBuf;
 
 #[derive(Serialize)]
-pub struct PluginItem {
+pub struct ToolInfo {
+    bins: Vec<ExecutableLocation>,
+    exe_path: PathBuf,
+    globals_dir: Option<PathBuf>,
+    globals_prefix: Option<String>,
     id: Id,
-    locator: PluginLocator,
+    inventory_dir: PathBuf,
+    metadata: ToolMetadataOutput,
     name: String,
-    version: Option<String>,
+    plugin: PluginLocator,
+    shims: Vec<ExecutableLocation>,
 }
 
 #[derive(Args, Clone, Debug)]
 pub struct ToolInfoArgs {
     #[arg(required = true, help = "ID of tool")]
     id: Id,
+
+    #[arg(long, help = "Print the info in JSON format")]
+    json: bool,
 }
 
 #[system]
@@ -27,6 +40,25 @@ pub async fn tool_info(args: ArgsRef<ToolInfoArgs>) {
     tool.resolve_version(&version, false).await?;
     tool.create_executables(false, false).await?;
     tool.locate_globals_dir().await?;
+
+    if args.json {
+        let info = ToolInfo {
+            bins: tool.get_bin_locations()?,
+            exe_path: tool.get_exe_path()?.to_path_buf(),
+            globals_dir: tool.get_globals_bin_dir().map(|p| p.to_path_buf()),
+            globals_prefix: tool.get_globals_prefix().map(|p| p.to_owned()),
+            inventory_dir: tool.get_inventory_dir(),
+            shims: tool.get_shim_locations()?,
+            id: tool.id,
+            name: tool.metadata.name.clone(),
+            metadata: tool.metadata,
+            plugin: tool.locator.unwrap(),
+        };
+
+        println!("{}", json::to_string_pretty(&info).into_diagnostic()?);
+
+        return Ok(());
+    }
 
     let mut printer = Printer::new();