+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 #### 🚀 Updates
 
+- Added a `proto tool info` command for viewing information about a tool and its plugin.
 - Added a `detect-strategy` setting to `~/.proto/config.toml` to configure which strategy to use when detecting a version. Accepts:
   - `first-available` (default) - Will use the first available version that is found. Either from `.prototools` or a tool specific file (`.nvmrc`, etc).
   - `prefer-prototools` - Prefer a `.prototools` version, even if found in a parent directory. If none found, falls back to tool specific file.